requirements.txt
fastapi
uvicorn
xmltodict
requests
apscheduler
unidecode
rapidfuzz
python-jose[cryptography]
python-multipart
jinja2
python-slugify
aiofiles
